Making proprietary schema definitions available for a reference.
