+++ src/proprietary-schemas.ts
@@ -1,0 +1,14 @@
+import { ZodDateIn } from "./date-in-schema";
+import { ZodDateOut } from "./date-out-schema";
+import { ZodFile } from "./file-schema";
+import { ZodUpload } from "./upload-schema";
+
+export type { ZodDateInDef } from "./date-in-schema";
+export type { ZodDateOutDef } from "./date-out-schema";
+export type { ZodFileDef } from "./file-schema";
+export type { ZodUploadDef } from "./upload-schema";
+
+export const file = ZodFile.create;
+export const upload = ZodUpload.create;
+export const dateIn = ZodDateIn.create;
+export const dateOut = ZodDateOut.create;
