+++ src/schema-walker.ts
@@ -1,5 +1,10 @@
 import { z } from "zod";
-import { ProprietaryKinds } from "./ez-namespace";
+import {
+  ZodDateInDef,
+  ZodDateOutDef,
+  ZodFileDef,
+  ZodUploadDef,
+} from "./proprietary-schemas";
 
 export type HandlingVariant = "last" | "regular" | "each";
 
@@ -28,6 +33,12 @@ export type SchemaHandler<
   Context extends object = {},
   Variant extends HandlingVariant = "regular"
 > = (params: SchemaHandlingProps<T, U, Context, Variant>) => U;
+
+export type ProprietaryKinds =
+  | ZodFileDef["typeName"]
+  | ZodUploadDef["typeName"]
+  | ZodDateInDef["typeName"]
+  | ZodDateOutDef["typeName"];
 
 export type HandlingRules<U, Context extends object = {}> = Partial<
   Record<
